// generated/src/app/api-models/smproductattribute-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmproductattributeS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  productattributeProductname: string;
  productattributeItfclabel: string;
  productattributeVersion: string;
  productattributeOs: string;
  productattributeProductsize: number;
  productattributePrdpath: string;
  productattributeParent: string;
  productattributeVisible: string;
  productattributeOpen: string;
  productattributeDescr: string;
}